requirements.txt
# Web Server
flask==2.3.0
flask-cors==4.0.0

# Voice Recognition & TTS
SpeechRecognition==3.10.0
PyAudio==0.2.13
pyttsx3==2.90

# Translation
googletrans==4.0.0rc1

# Sign Language Recognition
opencv-python==4.8.1.78
mediapipe==0.10.13
numpy==1.24.3
scikit-learn==1.3.2
joblib==1.4.2

# Optional: For dynamic sign language model
tensorflow-cpu==2.15.0

# Optional: For visualization
matplotlib==3.8.2